Move DiscoveryNetworksDeIE to Dplay
